秒杀功能实现

+++ src/main/java/com/seckill/mapper/OrderMapper.java
@@ -12,7 +12,7 @@ public interface OrderMapper {
 	@Insert("insert into order_info(user_id, goods_id, goods_name, goods_count, goods_price, order_channel, status, create_date)values("
 			+ "#{userId}, #{goodsId}, #{goodsName}, #{goodsCount}, #{goodsPrice}, #{orderChannel},#{status},#{createDate} )")
 	@SelectKey(keyColumn="id", keyProperty="id", resultType=long.class, before=false, statement="select last_insert_id()")
-	long insert(OrderInfo orderInfo);
+	int insert(OrderInfo orderInfo);
 
 	@Insert("insert into miaosha_order (user_id, goods_id, order_id)values(#{userId}, #{goodsId}, #{orderId})")
 	int insertMiaoshaOrder(MiaoshaOrder miaoshaOrder);
